Add configurable quick search option for filters
