+++ src/components/blocks/SearchFilters.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState, useMemo } from "react";
+import { ParsedUrlQuery } from "querystring";
 
 import useGetThemeConfig from "@hooks/useThemeConfig";
 
 import Tooltip from "@components/tooltip";
-import { ExternalLink } from "@components/ExternalLink";
 import { DateRange } from "../filters/DateRange";
 import { Accordian } from "@components/accordian/Accordian";
 import { InputListContainer } from "@components/filters/InputListContainer";
@@ -12,16 +12,17 @@ import { InputCheck } from "@components/forms/Checkbox";
 import { InputRadio } from "@components/forms/Radio";
 import { AppliedFilters } from "@components/filters/AppliedFilters";
 import Loader from "@components/Loader";
+import { FilterOptions } from "./FilterOptions";
 
 import { currentYear, minYear } from "@constants/timedate";
 import { QUERY_PARAMS } from "@constants/queryParams";
 
 import { getCountriesFromRegions } from "@helpers/getCountriesFromRegions";
 
-import { TGeography, TOrganisationDictionary, TSearchCriteria, TThemeConfig, TThemeConfigFilter, TThemeConfigOption } from "@types";
-import { ParsedUrlQuery } from "querystring";
 import { canDisplayFilter } from "@utils/canDisplayFilter";
 import { getFilterLabel } from "@utils/getFilterLabel";
+
+import { TGeography, TOrganisationDictionary, TSearchCriteria, TThemeConfigOption } from "@types";
 
 const { default: MethodologyLink } = await import(`/themes/${process.env.THEME}/components/MethodologyLink`);
 
@@ -35,102 +36,6 @@ const isCategoryChecked = (selectedCatgeory: string | undefined, themeConfigCate
   if (!selectedCatgeory && themeConfigCategory.slug.toLowerCase() === "all") return true;
 
   return false;
-};
-
-const getTaxonomyAllowedValues = (corporaKey: string, taxonomyKey: string, organisations: TOrganisationDictionary) => {
-  const allowedValues = organisations[corporaKey].corpora.find((corpus) => corpus.taxonomy.hasOwnProperty(taxonomyKey))?.taxonomy[taxonomyKey]
-    ?.allowed_values;
-
-  return allowedValues;
-};
-
-const renderFilterOptions = (
-  filter: TThemeConfigFilter,
-  query: ParsedUrlQuery,
-  handleFilterChange: Function,
-  organisations: TOrganisationDictionary,
-  themeConfig: TThemeConfig
-) => {
-  // If the filter has its own options defined, display them
-  if (filter.options && filter.options.length > 0) {
-    return filter.options.map((option) =>
-      filter.type === "radio" ? (
-        <InputRadio
-          key={option.slug}
-          label={option.label}
-          checked={query && query[QUERY_PARAMS[filter.taxonomyKey]] && query[QUERY_PARAMS[filter.taxonomyKey]].includes(option.slug)}
-          onChange={() => null} // supress normal radio behaviour to allow to deselection
-          onClick={() => {
-            handleFilterChange(QUERY_PARAMS[filter.taxonomyKey], option.slug, true);
-          }}
-        />
-      ) : (
-        <InputCheck
-          key={option.slug}
-          label={option.label}
-          checked={query && query[QUERY_PARAMS[filter.taxonomyKey]] && query[QUERY_PARAMS[filter.taxonomyKey]].includes(option.slug)}
-          onChange={() => {
-            handleFilterChange(QUERY_PARAMS[filter.taxonomyKey], option.slug);
-          }}
-        />
-      )
-    );
-  }
-  // Check the dependancy filter key for which filters to load the taxonomy for
-  let options = [];
-  const dependantFilter = themeConfig.filters.find((f) => f.taxonomyKey === filter.dependantFilterKey);
-  const queryDependantFilter = query[QUERY_PARAMS[dependantFilter?.taxonomyKey]] || [];
-
-  // If no filter of a given dependancy is selected, load all dependancy taxonomy values
-  if (queryDependantFilter.length === 0) {
-    for (let index = 0; index < dependantFilter.options.length; index++) {
-      const option = dependantFilter.options[index];
-      const taxonomyAllowedValues = getTaxonomyAllowedValues(option.corporaKey, filter.taxonomyKey, organisations);
-      options = options.concat(taxonomyAllowedValues);
-    }
-  } else {
-    // Otherwise, load the taxonomy values for the selected dependancy filter(s)
-    if (typeof queryDependantFilter === "string") {
-      const filterCoporaKey = dependantFilter.options.find((option) => option.slug === queryDependantFilter)?.corporaKey;
-      const taxonomyAllowedValues = getTaxonomyAllowedValues(filterCoporaKey, filter.taxonomyKey, organisations);
-      options = options.concat(taxonomyAllowedValues);
-    } else {
-      for (let index = 0; index < queryDependantFilter.length; index++) {
-        const filterCoporaKey = dependantFilter.options.find((option) => option.slug === queryDependantFilter[index])?.corporaKey;
-        const taxonomyAllowedValues = getTaxonomyAllowedValues(filterCoporaKey, filter.taxonomyKey, organisations);
-        options = options.concat(taxonomyAllowedValues);
-      }
-    }
-  }
-
-  // De-duplicate and sort the options
-  const optionsDeDuped: string[] = [...new Set(options.sort())];
-
-  if (optionsDeDuped.length) {
-    return optionsDeDuped.map((option: string) =>
-      filter.type === "radio" ? (
-        <InputRadio
-          key={option}
-          label={option}
-          checked={query && query[QUERY_PARAMS[filter.taxonomyKey]] && query[QUERY_PARAMS[filter.taxonomyKey]].includes(option)}
-          onChange={() => null} // supress normal radio behaviour to allow to deselection
-          onClick={() => {
-            handleFilterChange(QUERY_PARAMS[filter.taxonomyKey], option, true);
-          }}
-        />
-      ) : (
-        <InputCheck
-          key={option}
-          label={option}
-          checked={query && query[QUERY_PARAMS[filter.taxonomyKey]] && query[QUERY_PARAMS[filter.taxonomyKey]].includes(option)}
-          onChange={() => {
-            handleFilterChange(QUERY_PARAMS[filter.taxonomyKey], option);
-          }}
-        />
-      )
-    );
-  }
-  return null;
 };
 
 type TSearchFiltersProps = {
@@ -242,7 +147,15 @@ const SearchFilters = ({
               startOpen={filter.startOpen === "true" || !!query[QUERY_PARAMS[filter.taxonomyKey]]}
               showFade={filter.showFade}
             >
-              <InputListContainer>{renderFilterOptions(filter, query, handleFilterChange, organisations, themeConfig)}</InputListContainer>
+              <InputListContainer>
+                <FilterOptions
+                  filter={filter}
+                  query={query}
+                  handleFilterChange={handleFilterChange}
+                  organisations={organisations}
+                  themeConfig={themeConfig}
+                />
+              </InputListContainer>
             </Accordian>
           );
         })}
